Update keyField/data props Documentation (#1362)
I added more info to the descriptions of the keyField and data props of
the Table component. I also added an example of what these properties
might look like for a sample dataset.

I had trouble figuring these out, and had to find a medium post about
this library to understand how the input data should be formatted, and I
had to read the source code to understand what the keyField was for. I
think having better descriptions of these props will be helpful to
users.